+++ src/main/java/com/rafael/atendimento/controllers/ClassController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,7 +22,7 @@ import com.rafael.atendimento.dto.CalendarDTO;
 import com.rafael.atendimento.dto.ClassDTO;
 import com.rafael.atendimento.dto.ClassPageDTO;
 import com.rafael.atendimento.dto.ReportDTO;
-import com.rafael.atendimento.entity.Class;
+import com.rafael.atendimento.infra.security.TokenService;
 import com.rafael.atendimento.service.CalendarService;
 import com.rafael.atendimento.service.ClassService;
 import com.rafael.atendimento.service.ReportService;
@@ -40,6 +41,7 @@ public class ClassController {
 	private final ClassService classService;
 	private final CalendarService calendarService;
 	private final ReportService reportService;
+	private final TokenService tokenService;
 	
 //	@GetMapping
 //	public List<ClassDTO> findAll() {
@@ -47,20 +49,29 @@ public class ClassController {
 //	}
 	
 	@GetMapping
-    public ClassPageDTO list(@RequestParam(defaultValue = "0") @PositiveOrZero int page,
+    public ClassPageDTO list(
+    		@RequestHeader("Authorization") String token,
+    		@RequestParam(defaultValue = "0") @PositiveOrZero int page,
             @RequestParam(defaultValue = "10") @Positive @Max(100) int pageSize) {
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
         return classService.list(page, pageSize);
     }
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<ClassDTO> findById(@PathVariable Long id) {
+	public ResponseEntity<ClassDTO> findById(
+			@RequestHeader("Authorization") String token,
+			@PathVariable Long id) {
 		ClassDTO findClass = classService.findById(id);
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
 		return ResponseEntity.ok(findClass);
 	}
 	
 	@PostMapping
-    public ResponseEntity<?> createClass(@RequestBody @Valid ClassDTO classRequest) {
+    public ResponseEntity<?> createClass(
+    		@RequestHeader("Authorization") String token,
+    		@RequestBody @Valid ClassDTO classRequest) {
        try {
+    	   tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
     	   ClassDTO findClass = classService.create(classRequest);
     	   return ResponseEntity.ok(findClass);
        } catch (RuntimeException ex) {
@@ -69,8 +80,12 @@ public class ClassController {
     }
 	
 	@PutMapping("/{id}")
-    public ResponseEntity<?> updateClass(@PathVariable Long id, @RequestBody @Valid Class classRequest) {
+    public ResponseEntity<?> updateClass(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long id, 
+    		@RequestBody @Valid ClassDTO classRequest) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
 			ClassDTO findClass = classService.update(id, classRequest);
 			return ResponseEntity.ok(findClass);
 		} catch (RuntimeException ex) {
@@ -79,7 +94,10 @@ public class ClassController {
     }
 	
 	@DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteClass(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteClass(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long id) {
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
 		classService.delete(id);
         return ResponseEntity.noContent().build();
     }
@@ -87,8 +105,12 @@ public class ClassController {
 	// Métodos de Relacionamtentos
 	
 	@PostMapping("/{classId}/add-user/{userId}")
-	public ResponseEntity<?> addUserInClass(@PathVariable Long classId, @PathVariable Long userId) {
+	public ResponseEntity<?> addUserInClass(
+			@RequestHeader("Authorization") String token,
+			@PathVariable Long classId, 
+			@PathVariable Long userId) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), false);
 			ClassDTO updatedClass = classService.addUserInClass(classId, userId);
 			return ResponseEntity.ok(updatedClass);
         } catch (RuntimeException ex) {
@@ -97,8 +119,12 @@ public class ClassController {
 	}
 	
 	@DeleteMapping("/{classId}/remove-user/{userId}")
-	public ResponseEntity<?> removeUserInClass(@PathVariable Long classId, @PathVariable Long userId) {
+	public ResponseEntity<?> removeUserInClass(
+			@RequestHeader("Authorization") String token,
+			@PathVariable Long classId, 
+			@PathVariable Long userId) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), false);
 			ClassDTO updatedClass = classService.deleteUserInClass(classId, userId);
 			return ResponseEntity.ok(updatedClass);
         } catch (RuntimeException ex) {
@@ -107,8 +133,11 @@ public class ClassController {
 	}
 	
 	@GetMapping("/class-user/{userId}")
-    public ResponseEntity<?> getUserClasses(@PathVariable Long userId) {
+    public ResponseEntity<?> getUserClasses(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long userId) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
 			List<ClassDTO> classes = classService.getClassesForUser(userId);
 	        return ResponseEntity.ok(classes);
 		} catch (RuntimeException ex) {
@@ -118,8 +147,12 @@ public class ClassController {
 	
 	// Criar um novo calendário para a turma
 	@PostMapping("/{classId}/calendars")
-    public ResponseEntity<?> createCalendar(@PathVariable Long classId, @RequestBody @Valid CalendarDTO calendarDTO) {
+    public ResponseEntity<?> createCalendar(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId, 
+    		@RequestBody @Valid CalendarDTO calendarDTO) {
         try {
+        	tokenService.validateTokenAndPermissions(token, List.of("Admin", "Monitor", "Professor"), false);
             CalendarDTO newCalendar = calendarService.createCalendar(classId, calendarDTO);
             return ResponseEntity.status(HttpStatus.CREATED).body(newCalendar);
         } catch (RuntimeException ex) {
@@ -129,21 +162,32 @@ public class ClassController {
 	
 	// Listar todos os calendários de uma turma
     @GetMapping("/{classId}/calendars")
-    public ResponseEntity<List<CalendarDTO>> listCalendars(@PathVariable Long classId) {
+    public ResponseEntity<List<CalendarDTO>> listCalendars(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId) {
+    	tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
         List<CalendarDTO> calendars = calendarService.getCalendarsByClassId(classId);
         return ResponseEntity.ok(calendars);
     }
     
  // Deletar um calendário
     @DeleteMapping("/{classId}/calendars/{calendarId}")
-    public ResponseEntity<Void> deleteCalendar(@PathVariable Long classId, @PathVariable Long calendarId) {
+    public ResponseEntity<Void> deleteCalendar(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId, 
+    		@PathVariable Long calendarId) {
+    	tokenService.validateTokenAndPermissions(token, List.of("Admin", "Monitor", "Professor"), false);
         calendarService.deleteCalendar(classId, calendarId);
         return ResponseEntity.noContent().build();
     }
     
     @PutMapping("/{classId}/promote/{userId}")
-    public ResponseEntity<?> promoteToMonitor(@PathVariable Long classId, @PathVariable Long userId) {
+    public ResponseEntity<?> promoteToMonitor(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId, 
+    		@PathVariable Long userId) {
     	try {
+    		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
     		classService.promoteToMonitor(classId, userId);
     		return ResponseEntity.ok().build();
     	} catch (RuntimeException ex) {
@@ -152,8 +196,12 @@ public class ClassController {
     }
     
     @PutMapping("/{classId}/demote/{userId}")
-    public ResponseEntity<?> demoteToStudent(@PathVariable Long classId, @PathVariable Long userId) {
+    public ResponseEntity<?> demoteToStudent(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId, 
+    		@PathVariable Long userId) {
     	try {
+    		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
     		classService.demoteToStudent(classId, userId);
     		return ResponseEntity.ok().build();
     	} catch (RuntimeException ex) {
@@ -162,13 +210,19 @@ public class ClassController {
     }
     
     @GetMapping("/{classId}/report")
-    public ResponseEntity<ReportDTO> getClassReport(@PathVariable Long classId) {
+    public ResponseEntity<ReportDTO> getClassReport(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId) {
+    	tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), false);
         ReportDTO report = reportService.generateReport(classId);
         return ResponseEntity.ok(report);
     }
     
     @GetMapping("/{classId}/report/pdf")
-    public ResponseEntity<byte[]> generateReportPdf(@PathVariable Long classId) throws DocumentException, IOException {
+    public ResponseEntity<byte[]> generateReportPdf(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId) throws DocumentException, IOException {
+    	tokenService.validateTokenAndPermissions(token, List.of("Admin", "Monitor", "Professor"), false);
         byte[] pdfContent = reportService.generateReportPdf(classId);
         // Retornar o PDF como resposta
         HttpHeaders headers = new HttpHeaders();
