+++ src/main/java/com/rafael/atendimento/controllers/UserController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,6 +17,7 @@ import com.rafael.atendimento.dto.RegisterRequestDTO;
 import com.rafael.atendimento.dto.ResponseDTO;
 import com.rafael.atendimento.dto.UpdateRequestDTO;
 import com.rafael.atendimento.dto.UserDTO;
+import com.rafael.atendimento.infra.security.TokenService;
 import com.rafael.atendimento.service.UserService;
 
 import lombok.RequiredArgsConstructor;
@@ -26,21 +28,29 @@ import lombok.RequiredArgsConstructor;
 public class UserController {
 	
 	private final UserService userService;
+	private final TokenService tokenService;
 	
 	@GetMapping
-	public List<UserDTO> findAll() {
+	public List<UserDTO> findAll(@RequestHeader("Authorization") String token) {
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), false);
 		return userService.findAll();
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<UserDTO> findById(@PathVariable Long id) {
+	public ResponseEntity<UserDTO> findById(
+			@RequestHeader("Authorization") String token, 
+			@PathVariable Long id) {
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
 		UserDTO user = userService.findById(id);
 		return ResponseEntity.ok(user);
 	}
 	
 	@PostMapping
-    public ResponseEntity<?> createUser(@RequestBody RegisterRequestDTO userRequest) {
+    public ResponseEntity<?> createUser(
+    		@RequestHeader("Authorization") String token, 
+    		@RequestBody RegisterRequestDTO userRequest) {
        try {
+    	   tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
     	   UserDTO user = userService.create(userRequest);
     	   return ResponseEntity.ok(user);
        } catch (RuntimeException ex) {
@@ -49,20 +59,30 @@ public class UserController {
     }
 	
 	@PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable Long id, @RequestBody UpdateRequestDTO userRequest) {
-	   ResponseDTO user = userService.update(id, userRequest);
-	   return ResponseEntity.ok(user);
+    public ResponseEntity<?> updateUser(
+    		@RequestHeader("Authorization") String token, 
+    		@PathVariable Long id, 
+    		@RequestBody UpdateRequestDTO userRequest) {
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
+		ResponseDTO user = userService.update(id, userRequest);
+		return ResponseEntity.ok(user);
     }
 	
 	@DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUser(
+    		@RequestHeader("Authorization") String token, 
+    		@PathVariable Long id) {
+		tokenService.validateTokenAndPermissions(token, List.of("Admin", "Professor"), false);
 		userService.delete(id);
         return ResponseEntity.noContent().build();
     }
 	
 	@GetMapping("/user-class/{classId}")
-    public ResponseEntity<?> getUserClasses(@PathVariable Long classId) {
+    public ResponseEntity<?> getUserClasses(
+    		@RequestHeader("Authorization") String token, 
+    		@PathVariable Long classId) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
 			List<UserDTO> users = userService.getAllStudentsInClass(classId);
 	        return ResponseEntity.ok(users);
 		} catch (RuntimeException ex) {
@@ -71,8 +91,11 @@ public class UserController {
     }
 	
 	@GetMapping("/userTeachers-class/{classId}")
-    public ResponseEntity<?> getUserTeachersClasses(@PathVariable Long classId) {
+    public ResponseEntity<?> getUserTeachersClasses(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
 			List<UserDTO> users = userService.getAllTeachersInClass(classId);
 	        return ResponseEntity.ok(users);
 		} catch (RuntimeException ex) {
@@ -81,8 +104,11 @@ public class UserController {
     }
 	
 	@GetMapping("/userMonitors-class/{classId}")
-    public ResponseEntity<?> getUserMonitorsClasses(@PathVariable Long classId) {
+    public ResponseEntity<?> getUserMonitorsClasses(
+    		@RequestHeader("Authorization") String token,
+    		@PathVariable Long classId) {
 		try {
+			tokenService.validateTokenAndPermissions(token, List.of("Admin", "Aluno", "Monitor", "Professor"), true);
 			List<UserDTO> users = userService.getAllMonitorsInClass(classId);
 	        return ResponseEntity.ok(users);
 		} catch (RuntimeException ex) {
